third_party/WebKit/LayoutTests/fast/dom/shadow/outer-scope-lower-specificity-wins.html

<!DOCTYPE html>
<script src="../../../resources/js-test.js"></script>
<script src="resources/shadow-dom.js"></script>
<div id="sandbox"></div>
<script>
description('Inner scope rules wins, with higher specificity (v0 bug compatibility).');

sandbox.appendChild(
    createDOM('div', {'id': 'host'},
        createShadowRoot(
            createDOM('style', {},
                document.createTextNode(
                    '::content #target { color: red; }')),
            createDOM('content', {})),
        createDOM('style', {},
            document.createTextNode(
                'span { color: green; }')),
        createDOM('span', {'id': 'target'})));

shouldBeEqualToString('getComputedStyle(target).color', 'rgb(255, 0, 0)');
</script>
