chromium/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/outer-scope-lower-specificity-wins-expected.txt

Inner scope rules wins, with higher specificity (v0 bug compatibility).

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL getComputedStyle(target).color should be rgb(255, 0, 0). Was rgb(0, 128, 0).
PASS successfullyParsed is true

TEST COMPLETE
